refactor: move clipboard server impls to their own subfolder
