+++ vicinae/CMakeLists.txt
@@ -143,10 +143,10 @@ set(SRCS
 	src/services/clipboard/clipboard-server.hpp
 	src/services/clipboard/clipboard-db.hpp
 	src/services/clipboard/clipboard-db.cpp
-	src/services/clipboard/wlr-clipboard-server.hpp
-	src/services/clipboard/wlr-clipboard-server.cpp
-	src/services/clipboard/dummy-clipboard-server.hpp
-	src/services/clipboard/dummy-clipboard-server.cpp
+	src/services/clipboard/wlr/wlr-clipboard-server.hpp
+	src/services/clipboard/wlr/wlr-clipboard-server.cpp
+	src/services/clipboard/dummy/dummy-clipboard-server.hpp
+	src/services/clipboard/dummy/dummy-clipboard-server.cpp
 	src/services/clipboard/clipboard-server-factory.hpp
 	src/services/clipboard/clipboard-server-factory.cpp
 	src/services/clipboard/clipboard-service.hpp
