+++ vicinae/src/services/clipboard/dummy/dummy-clipboard-server.hpp
